post-common-core/cyber/piscine/00-arachnida/scorpion/requirements.txt
attrs==25.4.0
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
exceptiongroup==1.3.0
flake8==7.3.0
h11==0.16.0
idna==3.11
mccabe==0.7.0
outcome==1.3.0.post0
pillow==12.0.0
pycodestyle==2.14.0
pyflakes==3.4.0
PySocks==1.7.1
requests==2.32.5
selenium==4.38.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.5.0
websocket-client==1.9.0
wsproto==1.2.0
